Add update user

+++ api/user/user.controller.js
@@ -1,4 +1,5 @@
 import { loggerService } from "../../services/logger.service.js"
+import { reviewService } from "../review/review.service.js"
 import { userService } from "./user.service.js"
 
 
@@ -26,10 +27,35 @@ export async function getUserById(req, res) {
 export async function removeUser(req, res) {
     try {
         const userId = req.params.id
+
+        const isUserHaveReviews = await reviewService.query({ byUserId: userId })
+        if (isUserHaveReviews?.length > 0) {
+            return res.status(400).send('Cannot remove user with reviews.')
+        }
+
         await userService.remove(userId)
         res.send('User removed')
     } catch (err) {
         loggerService.error('cannot remove user', err)
         res.status(500).send('cannot remove user')
+    }
+}
+
+export async function updateUser(req, res) {
+    const { loggedinUser, body: user } = req
+    console.log('loggedinUser:', loggedinUser)
+    console.log('user:', user)
+
+
+    if (loggedinUser?._id !== user?._id) {
+        return res.status(400).send('You are not authorized')
+    }
+
+    try {
+        const updatedUser = await userService.update(user)
+        res.send(updatedUser)
+    } catch (err) {
+        loggerService.error('cannot update user', err)
+        res.status(500).send('cannot update users')
     }
 }
